Updates to add a `attributesExcluded` param;
I have finished the basic updates for an exclusion array passed in, but I haven't yet been able to work on the tests inside of the normalizer package. It passes tests in my app, and doesn't break the existing tests in any way.

When creating a new normalizer you would now be able to pass in an `attributesExcluded` option with an array of attributes that will be blacklisted during normalization.

Example: ```
var normalizer = new Normalizer({ "attributes": null, "attributesExcluded": ['data-a']});
```

Please let me know if you have any questions or concerns. Feel free to make this idea whatever you want. :smile:

(P.S.: If you have time to help me with the tests, that's great. I have a busy day tomorrow.)